allow aa mutations for insertions with raw reads.
assuming deletions are always provided as the nucleotides. With substitutions both the codon_pos and nuc_pos will be the codon position so to avoid the edge case where a nuc mutation is found on the nuc position corresponding to the codon position an additional if statement where made for substitutions only looking for the AA mutation at the codon position.